+++ changelog.md
@@ -3,7 +3,7 @@
 ## NEXT
 
 - Fixed a bug in the `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.get_topological_order` method in
-  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` where the method was not working properly.
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph`.
 
 ## 0.3.8